requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
psycopg==3.2.4
gunicorn==23.0.0
python-decouple==3.8
drf-spectacular==0.28.0
django-bootstrap5==25.1